requirements.txt
altgraph==0.17.3
cffi==1.15.1
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
configobj==5.0.8
configparser==5.3.0
cryptography==41.0.1
etelemetry==0.3.0
filelock==3.12.2
fitz==0.0.1.dev2
httplib2==0.22.0
isodate==0.6.1
looseversion==1.2.0
lxml==4.9.2
macholib==1.16.2
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.0
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
prov==2.0.0
pycparser==2.21
pydot==1.4.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
pyparsing==3.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
pyxnat==1.5
rdflib==6.3.2
scipy==1.11.1
simplejson==3.19.1
six==1.16.0
tesserocr==2.6.0
traits==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
Wand==0.6.11
